Proof of concept sample selection, based on tuning phase
This is based on an old spreadsheet that's been circulating for a while. The idea is to choose a looping DPCM sample length that most closely corresponds with some multiple of a target MIDI frequency. By generating a sample that fills this length exactly, that frequency can be cleanly approximated with a small sample size.