Implement safe_amplitude
Basically, above a certain frequency for the fundamental, the 1-bit delta encoding can't keep up and falls behind. We can't let this happen for a looping sample, otherwise the end of the sample is out of phase (because it hasn't caught up), so we need to detect this case and compensate for it.