Artificial "high/low" fundamental, for generating pure triangles
It is occasionally useful to generate samples which force the DPCM
counter to emit a string of 1 bits, followed by a string of 0 bits.
This creates an exceptionally clean triangle, at the cost of a fixed
amplitude and difficulty tuning. This generator allows automating
this process somewhat. For decent results, generate a high pitched
triangle (above C4 or so) and then downpitch the samples to fill
out the mid and bass, so the amplitude is consistent(ish).

Can also be used in the sub-bass range to generate a *very* loud
trapezoid, which might be useful.