Actually respect max length. Do NOT start DPCM at level 0! (Bugfix)
The DPCM level thing was the root cause behind unclean loops,
especially when using waveform type samples, which are typically
centered around 64 or so. The routine was *supposed* to be starting
the PCM level chaser around the first incoming sample, to avoid it
racing away from "center" near loop transitions, but for some reason
I stupidly passed in an explicit 0, disabling the feature. Whoops!