Specify "python3.8" as the desired virtualenv
This file is meant for use with pyenv-virtualenv. Specifically,
after that's configured, one can run:

  pyenv virtualenv 3.8.12 python3.8

This will install a testing copy of python 3.8.12 and a virtualenv
for packages. Then that virtualenv will activate when switching to
this project directory, and deactivate when switching elsewhere.
Convenient!

Note that this project will be standardizing on Python 3.8 for the
time being, as Windows 7 users aren't able to install anything
newer. (I will not be spending effort trying to support older,
end of life Windows versions, but 7 is still pretty common in the
wild, and it's not EOL until 2023)